Move collect commands under collect category
