+++ src/cli/commands/collect/local.ts
@@ -9,14 +9,14 @@ import {
   resolveActrunHeadBranch,
   resolveActrunStartedAt,
   type ActrunRunOutput,
-} from "../adapters/actrun.js";
-import { playwrightAdapter } from "../adapters/playwright.js";
-import { junitAdapter } from "../adapters/junit.js";
-import type { MetricStore, WorkflowRun, TestResult } from "../storage/types.js";
-import { toStoredTestResult } from "../storage/test-result-mapper.js";
-import { collectCommitChanges } from "./collect-commit-changes.js";
-import { resolveCurrentCommitSha } from "../core/git.js";
-import { exportRunParquet } from "./export-parquet.js";
+} from "../../adapters/actrun.js";
+import { playwrightAdapter } from "../../adapters/playwright.js";
+import { junitAdapter } from "../../adapters/junit.js";
+import type { MetricStore, WorkflowRun, TestResult } from "../../storage/types.js";
+import { toStoredTestResult } from "../../storage/test-result-mapper.js";
+import { collectCommitChanges } from "./commit-changes.js";
+import { resolveCurrentCommitSha } from "../../core/git.js";
+import { exportRunParquet } from "../export-parquet.js";
 
 export interface CollectLocalOpts {
   store: MetricStore;
